feat admin can delete user + see all users

+++ front/src/components/UserProfile/DeleteProfile.vue
@@ -4,7 +4,7 @@
       <h4 slot="header" class="card-title">Delete profile</h4>
       <p class="text-container">
         Warning ! Clicking on this button will definitly delete all of your
-        personnal and work informations.
+        personnal and work information.
       </p>
       <div class="text-center">
         <button
@@ -20,28 +20,35 @@
 </template>
 
 <script>
-import Cookies from 'js-cookie';
-import jwt_decode from 'jwt-decode';
-import { getConnexionType } from '../../helpers/getConnexionType'
+import Cookies from "js-cookie";
+import jwt_decode from "jwt-decode";
+import { getConnexionType } from "../../helpers/getConnexionType";
 
 export default {
-  name: 'DeleteProfile',
+  name: "DeleteProfile",
   props: {
     deleteUser: {
       type: Function,
-      required: true,
-    },
+      required: true
+    }
   },
   methods: {
     async handleDeleteUser() {
-      if (getConnexionType() === 'none') {
-        alert("Oops ! You must be connected to the Internet to use this feature");
+      if (getConnexionType() === "none") {
+        alert(
+          "Oops ! You must be connected to the Internet to use this feature"
+        );
         return;
       }
-      const token = Cookies.get('token');
-      this.deleteUser(jwt_decode(token).id);
-    },
-  },
+      const token = Cookies.get("token");
+      if (confirm("Do you really want to delete your profile ?")) {
+        this.deleteUser(jwt_decode(token).id);
+        Cookies.remove("token");
+        window.localStorage.clear();
+        this.$router.push("signup");
+      }
+    }
+  }
 };
 </script>
